Testing: testing code

+++ classes/local/unassignment_process/unassignments/unassignment_controller.php
@@ -82,11 +82,13 @@ class unassignment_controller {
      * @return void
      */
     public function process_unassignments(): void {
-        foreach ($this->allaffectedunits as $unitid) {
-            foreach ($this->allaffectedusers as $userid) {
-                $this->unitmemberrepository->remove($userid, $unitid);
+        if (!empty($this->allaffectedusers)) {
+            foreach ($this->allaffectedunits as $unitid) {
+                foreach ($this->allaffectedusers as $userid) {
+                    $this->unitmemberrepository->remove($userid, $unitid);
+                }
+                assignments_facade::delete_assignments($this->allaffectedrules, $userid);
             }
-            assignments_facade::delete_assignments($this->allaffectedrules, $userid);
         }
     }
